refactor: templateコンポーネントは他のtemplateコンポーネントに依存しない
