+++ components/templates/Books.tsx
@@ -5,16 +5,12 @@ import Button from "@material-ui/core/Button";
 import Container from "@material-ui/core/Container";
 import Typography from "@material-ui/core/Typography";
 import AddIcon from "@material-ui/icons/Add";
-import Book from "$templates/Book";
 import ActionHeader from "$organisms/ActionHeader";
 import BookPreview from "$organisms/BookPreview";
-import BookPreviewDialog from "$organisms/BookPreviewDialog";
 import SortSelect from "$atoms/SortSelect";
 import CreatorFilter from "$atoms/CreatorFilter";
 import SearchTextField from "$atoms/SearchTextField";
-import useDialogProps from "$utils/useDialogProps";
 import type { BookSchema } from "$server/models/book";
-import type { TopicSchema } from "$server/models/topic";
 import { SortOrder } from "$server/models/sortOrder";
 import { Filter } from "$types/filter";
 import useContainerStyles from "styles/container";
@@ -43,9 +39,7 @@ export type Props = {
   onLoadMore?(): void;
   onBookClick?(book: BookSchema): void;
   onBookEditClick(book: BookSchema): void;
-  onBookLinkClick(book: BookSchema): void;
   onBookNewClick(): void;
-  onTopicEditClick(topic: TopicSchema): void;
   onSortChange?(sort: SortOrder): void;
   onFilterChange?(filter: Filter): void;
 };
@@ -57,21 +51,14 @@ export default function Books(props: Props) {
     loading = false,
     hasNextPage = false,
     onLoadMore = () => undefined,
+    onBookClick,
     onBookEditClick,
-    onBookLinkClick,
     onBookNewClick,
-    onTopicEditClick,
     onSortChange,
     onFilterChange,
   } = props;
   const { query, onSearchInput, onLtiContextClick } = useSearchAtom();
-  const {
-    data: dialog,
-    open,
-    onClose,
-    dispatch,
-  } = useDialogProps<BookSchema>();
-  const handleBookClick = (book: BookSchema) => () => dispatch(book);
+  const handleBookClick = (book: BookSchema) => () => onBookClick?.(book);
   const handleBookEditClick = (book: BookSchema) => () => onBookEditClick(book);
   const handleBookNewClick = () => onBookNewClick();
   const classes = useStyles();
@@ -143,19 +130,6 @@ export default function Books(props: Props) {
             [...Array(5)].map((_, i) => <Skeleton key={i} height={64} />)}
         </div>
       </Container>
-      {dialog && (
-        <BookPreviewDialog open={open} onClose={onClose} book={dialog}>
-          {(props) => (
-            <Book
-              {...props}
-              onBookEditClick={onBookEditClick}
-              onBookLinkClick={onBookLinkClick}
-              onOtherBookLinkClick={onClose}
-              onTopicEditClick={onTopicEditClick}
-            />
-          )}
-        </BookPreviewDialog>
-      )}
     </div>
   );
 }
